+++ csharp/src/Ice/ReplyStatus.cs
@@ -1,0 +1,19 @@
+//
+// Copyright (c) ZeroC, Inc. All rights reserved.
+//
+
+namespace Ice
+{
+    /// <summary>The reply status used in Ice1 response frames (known as reply messages).</summary>
+    public enum ReplyStatus : byte
+    {
+        OK = 0,
+        UserException = 1,
+        ObjectNotExistException = 2,
+        FacetNotExistException = 3,
+        OperationNotExistException = 4,
+        UnknownLocalException = 5,
+        UnknownUserException = 6,
+        UnknownException = 7
+    }
+}
