+++ csharp/src/Ice/ResponseFrame.cs
@@ -11,28 +11,16 @@ namespace Ice
 {
     using Context = Dictionary<string, string>;
 
-    /// <summary>The reply status used in Ice1 response frames (known as reply messages).</summary>
-    public enum ReplyStatus : byte
-    {
-        OK = 0,
-        UserException = 1,
-        ObjectNotExistException = 2,
-        FacetNotExistException = 3,
-        OperationNotExistException = 4,
-        UnknownLocalException = 5,
-        UnknownUserException = 6,
-        UnknownException = 7
-    }
-
     /// <summary>Represents a response protocol frame received by the application.</summary>
     // TODO: IncomingResponseFrame should derive from InputStream
     public sealed class IncomingResponseFrame
     {
-        // TODO: missing RequestId ?
+        // TODO: missing RequestId, Connection, Encoding?
 
         public InputStream InputStream { get; }
 
         /// <summary>The Ice1 reply status. Only meaningful for the Ice1 protocol, always set to OK with Ice2.</summary>
+        // TODO: currently it's only OK or UserException as we "throw" other responses.
         public ReplyStatus ReplyStatus { get; }
 
         /// <summary>The response context. Always null with Ice1.</summary>
@@ -40,19 +28,22 @@ namespace Ice
 
         /// <summary>The payload of this response frame. The bytes inside the payload should not be written to;
         /// they are writable because of the <see cref="System.Net.Sockets.Socket"/> methods for sending.</summary>
-        // TODO: describe how long this payload remains valid once we add memory pooling.
         public ArraySegment<byte> Payload
         {
             get
             {
                 if (_payload == null)
                 {
-                    if (InputStream.IsEmpty) // TODO, it should never be empty, but currently it is when fulfilled by Sent.
+                    // TODO, it should never be empty, but currently it is when fulfilled by Sent
+                    if (InputStream.IsEmpty)
                     {
                         _payload = ArraySegment<byte>.Empty;
                     }
                     else
                     {
+                        // TODO: for other reply status,return the non-encaps payload.
+                        Debug.Assert(ReplyStatus == ReplyStatus.OK || ReplyStatus == ReplyStatus.UserException);
+
                         // TODO: works only when Payload called first before reading anything. Need a better version!
                         // TODO: not efficient to create an array here
                         // TODO: provide Encoding property
